backend/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
matplotlib>=3.4.2
seaborn>=0.11.1
joblib>=1.0.1
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.26.0
python-multipart>=0.0.5
pydantic>=1.8.2
nltk>=3.8.0
tqdm>=4.65.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
aiohttp>=3.8.0
async-timeout>=4.0.0 
protobuf>=3.20.3,<5.0.0
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0
pytest-mock>=3.6.1
aioresponses>=0.7.2
seaborn>=0.11.1
